完成代码改造

+++ zzyl-nursing-platform/src/main/java/com/zzyl/nursing/domain/NursingLevel.java
@@ -1,5 +1,6 @@
 package com.zzyl.nursing.domain;
 
+import java.math.BigDecimal;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 import org.apache.commons.lang3.builder.ToStringStyle;
 import com.zzyl.common.annotation.Excel;
@@ -8,7 +9,7 @@ import com.zzyl.common.core.domain.BaseEntity;
 /**
  * 护理等级对象 nursing_level
  * 
- * @author ruoyi
+ * @author hansheep
  * @date 2026-02-08
  */
 public class NursingLevel extends BaseEntity
@@ -28,7 +29,7 @@ public class NursingLevel extends BaseEntity
 
     /** 护理费用 */
     @Excel(name = "护理费用")
-    private Double fee;
+    private BigDecimal fee;
 
     /** 状态（0：禁用，1：启用） */
     @Excel(name = "状态", readConverterExp = "0=：禁用，1：启用")
@@ -47,6 +48,7 @@ public class NursingLevel extends BaseEntity
     {
         return id;
     }
+
     public void setName(String name) 
     {
         this.name = name;
@@ -56,6 +58,7 @@ public class NursingLevel extends BaseEntity
     {
         return name;
     }
+
     public void setLplanId(Long lplanId) 
     {
         this.lplanId = lplanId;
@@ -65,15 +68,17 @@ public class NursingLevel extends BaseEntity
     {
         return lplanId;
     }
-    public void setFee(Double fee) 
+
+    public void setFee(BigDecimal fee) 
     {
         this.fee = fee;
     }
 
-    public Double getFee() 
+    public BigDecimal getFee() 
     {
         return fee;
     }
+
     public void setStatus(Integer status) 
     {
         this.status = status;
@@ -83,6 +88,7 @@ public class NursingLevel extends BaseEntity
     {
         return status;
     }
+
     public void setDescription(String description) 
     {
         this.description = description;
